Store ErlNifPid instead of ERL_NIF_TERM
Storing an ErlNifPid does not require an independant
environment as an ERL_NIF_TERM does. Since we're only
really interested in using this for process pooling,
the limiting functionality is justified.

CQueue will benefit the most from this, as members of
its queues frequently leave and then rejoin, resulting
in the allocation and freeing of a lot of environments
during standard calls. No more.